Refactor rainfall forcing into simulation loop
