+++ src/nefas/engine.py
@@ -1,0 +1,287 @@
+from __future__ import annotations
+
+import logging
+from dataclasses import dataclass
+from pathlib import Path
+
+import geopandas as gpd
+import matplotlib
+import numpy as np
+import rasterio
+from rasterio.features import rasterize
+
+matplotlib.use("Agg")
+from matplotlib import pyplot as plt
+
+from .config import RainfallConfig, SimulationConfig
+from .preprocessing import IntermediateOutputs
+from .simulation import RasterGrid, SimulationState
+
+LOGGER = logging.getLogger(__name__)
+
+
+@dataclass(frozen=True)
+class SimulationRunResult:
+    """Files and timing produced by a simulation run."""
+
+    snapshot_directory: Path
+    snapshots: tuple[Path, ...]
+    duration_seconds: float
+
+
+def run_simulation(
+    config: SimulationConfig,
+    intermediates: IntermediateOutputs,
+) -> SimulationRunResult:
+    """Run the simulation loop and write PNG snapshots."""
+    grid = RasterGrid.from_dem(intermediates.clipped_dem)
+    state = SimulationState.dry(grid, manning_n=0.08, runoff_coefficient=1.0)
+    storm_mask = storm_footprint_mask(config.inputs.storm_footprint, intermediates.clipped_dem)
+    duration_seconds = rainfall_duration_seconds(config.rainfall)
+    snapshot_interval_seconds = config.output.snapshots.interval_minutes * 60
+
+    snapshot_directory = resolve_snapshot_directory(
+        config.output.snapshots.directory,
+        intermediates.workspace,
+    )
+    snapshot_directory.mkdir(parents=True, exist_ok=True)
+
+    snapshots = [
+        write_snapshot(
+            grid,
+            storm_mask,
+            state,
+            snapshot_directory,
+            index=0,
+        )
+    ]
+    next_snapshot_seconds = min(snapshot_interval_seconds, duration_seconds)
+    while state.hydraulic.time_seconds < duration_seconds:
+        dt_seconds = timestep_duration_seconds(
+            current_time_seconds=state.hydraulic.time_seconds,
+            duration_seconds=duration_seconds,
+            next_snapshot_seconds=next_snapshot_seconds,
+            nominal_time_step_seconds=config.time_step.seconds,
+        )
+        run_timestep(
+            state,
+            config.rainfall,
+            storm_mask,
+            dt_seconds,
+        )
+
+        if state.hydraulic.time_seconds >= next_snapshot_seconds:
+            snapshots.append(
+                write_snapshot(
+                    grid,
+                    storm_mask,
+                    state,
+                    snapshot_directory,
+                    index=len(snapshots),
+                )
+            )
+            next_snapshot_seconds = min(
+                next_snapshot_seconds + snapshot_interval_seconds,
+                duration_seconds,
+            )
+
+    LOGGER.info("Wrote %s simulation snapshots to %s", len(snapshots), snapshot_directory)
+    return SimulationRunResult(
+        snapshot_directory=snapshot_directory,
+        snapshots=tuple(snapshots),
+        duration_seconds=state.hydraulic.time_seconds,
+    )
+
+
+def resolve_snapshot_directory(snapshot_directory: Path, workspace: Path) -> Path:
+    """Resolve relative snapshot directories beneath the run workspace."""
+    if snapshot_directory.is_absolute():
+        return snapshot_directory
+    return workspace / snapshot_directory
+
+
+def rainfall_duration_seconds(rainfall: RainfallConfig) -> float:
+    """Return the duration implied by the rainfall series."""
+    return max(point.time_minutes for point in rainfall.series) * 60
+
+
+def timestep_duration_seconds(
+    current_time_seconds: float,
+    duration_seconds: float,
+    next_snapshot_seconds: float,
+    nominal_time_step_seconds: float,
+) -> float:
+    """Return the next timestep without stepping past a snapshot or event end."""
+    return min(
+        nominal_time_step_seconds,
+        duration_seconds - current_time_seconds,
+        next_snapshot_seconds - current_time_seconds,
+    )
+
+
+def storm_footprint_mask(storm_footprint: Path, template_raster: Path) -> np.ndarray:
+    """Rasterize the storm footprint onto the clipped DEM grid."""
+    storm = gpd.read_file(storm_footprint)
+    if storm.empty:
+        raise RuntimeError("Storm footprint does not contain any features.")
+
+    with rasterio.open(template_raster) as template:
+        storm = storm.to_crs(template.crs)
+        geometries = [geometry for geometry in storm.geometry if geometry is not None and not geometry.is_empty]
+        if not geometries:
+            raise RuntimeError("Storm footprint does not contain any valid geometries.")
+
+        mask = rasterize(
+            ((geometry, 1) for geometry in geometries),
+            out_shape=(template.height, template.width),
+            transform=template.transform,
+            fill=0,
+            dtype="uint8",
+            all_touched=True,
+        ).astype(bool)
+        valid_cells = template.read_masks(1) > 0
+
+    mask &= valid_cells
+    if not mask.any():
+        raise RuntimeError("Storm footprint does not overlap the clipped DEM.")
+    return mask
+
+
+def run_timestep(
+    state: SimulationState,
+    rainfall: RainfallConfig,
+    storm_mask: np.ndarray,
+    dt_seconds: float,
+) -> None:
+    """Advance the full model by one timestep."""
+    timestep_start_seconds = state.hydraulic.time_seconds
+    apply_rainfall_forcing(state, rainfall, storm_mask, timestep_start_seconds, dt_seconds)
+    water_timestep(state, dt_seconds)
+    state.hydraulic.time_seconds += dt_seconds
+    update_diagnostics(state, dt_seconds)
+
+
+def apply_rainfall_forcing(
+    state: SimulationState,
+    rainfall: RainfallConfig,
+    storm_mask: np.ndarray,
+    timestep_start_seconds: float,
+    dt_seconds: float,
+) -> None:
+    """Add rainfall depth for one timestep when the event is active."""
+    rate_m_per_second = average_rainfall_rate_m_per_second(
+        rainfall,
+        timestep_start_seconds,
+        dt_seconds,
+    )
+    if rate_m_per_second <= 0:
+        return
+
+    add_rainfall_depth(state, storm_mask, rate_m_per_second, dt_seconds)
+
+
+def average_rainfall_rate_m_per_second(
+    rainfall: RainfallConfig,
+    timestep_start_seconds: float,
+    dt_seconds: float,
+) -> float:
+    """Return the mean rainfall rate across a timestep."""
+    start_rate = rainfall_rate_m_per_second(rainfall, timestep_start_seconds)
+    end_rate = rainfall_rate_m_per_second(rainfall, timestep_start_seconds + dt_seconds)
+    return 0.5 * (start_rate + end_rate)
+
+
+def rainfall_rate_m_per_second(rainfall: RainfallConfig, elapsed_seconds: float) -> float:
+    """Linearly interpolate rainfall intensity and return meters per second."""
+    elapsed_minutes = elapsed_seconds / 60
+    points = rainfall.series
+
+    if elapsed_minutes <= points[0].time_minutes:
+        return points[0].rate_mm_per_hr / 1000 / 3600
+
+    for previous, current in zip(points, points[1:]):
+        if elapsed_minutes <= current.time_minutes:
+            elapsed_fraction = (
+                (elapsed_minutes - previous.time_minutes)
+                / (current.time_minutes - previous.time_minutes)
+            )
+            rate_mm_per_hr = previous.rate_mm_per_hr + (
+                (current.rate_mm_per_hr - previous.rate_mm_per_hr) * elapsed_fraction
+            )
+            return rate_mm_per_hr / 1000 / 3600
+
+    return points[-1].rate_mm_per_hr / 1000 / 3600
+
+
+def add_rainfall_depth(
+    state: SimulationState,
+    storm_mask: np.ndarray,
+    rate_m_per_second: float,
+    dt_seconds: float,
+) -> None:
+    """Add rainfall depth to valid cells covered by the storm footprint."""
+    receiving_cells = storm_mask & state.grid.valid_cells
+    depth_delta = rate_m_per_second * dt_seconds
+    state.hydraulic.depth[receiving_cells] += (
+        depth_delta * state.surface.runoff_coefficient[receiving_cells]
+    )
+
+
+def water_timestep(state: SimulationState, dt_seconds: float) -> None:
+    """Move water during one timestep once the finite-volume solver exists."""
+    pass
+
+
+def update_diagnostics(state: SimulationState, dt_seconds: float) -> None:
+    """Update diagnostic rasters after a timestep has been applied."""
+    state.diagnostics.max_depth = np.maximum(
+        state.diagnostics.max_depth,
+        state.hydraulic.depth,
+    )
+    wet_cells = state.hydraulic.depth > 0
+    newly_wet = np.isnan(state.diagnostics.arrival_time) & wet_cells
+    state.diagnostics.arrival_time[newly_wet] = state.hydraulic.time_seconds
+    state.diagnostics.flood_duration[wet_cells] += dt_seconds
+
+
+def write_snapshot(
+    grid: RasterGrid,
+    storm_mask: np.ndarray,
+    state: SimulationState,
+    snapshot_directory: Path,
+    index: int,
+) -> Path:
+    """Write one simulation snapshot and return its path."""
+    snapshot = snapshot_directory / f"snapshot_{index:04d}.png"
+    render_snapshot(
+        grid,
+        storm_mask,
+        state.hydraulic.depth,
+        snapshot,
+        elapsed_minutes=state.hydraulic.time_seconds / 60,
+    )
+    return snapshot
+
+
+def render_snapshot(
+    grid: RasterGrid,
+    storm_mask: np.ndarray,
+    depth: np.ndarray,
+    path: Path,
+    elapsed_minutes: float,
+) -> None:
+    """Render DEM, storm footprint, and water depth into a PNG image."""
+    figure, axis = plt.subplots(figsize=(8, 6), dpi=150)
+    axis.imshow(np.ma.masked_invalid(grid.elevation), cmap="gray")
+    axis.imshow(np.where(storm_mask, 1.0, np.nan), cmap="Blues", alpha=0.22, vmin=0, vmax=1)
+
+    depth_layer = np.ma.masked_where(depth <= 0, depth)
+    water = axis.imshow(depth_layer, cmap="turbo", alpha=0.75)
+    if np.nanmax(depth) > 0:
+        colorbar = figure.colorbar(water, ax=axis, fraction=0.046, pad=0.04)
+        colorbar.set_label("Water depth (m)")
+
+    axis.set_title(f"Simulated inundation at {elapsed_minutes:.1f} minutes")
+    axis.axis("off")
+    figure.savefig(path, bbox_inches="tight")
+    plt.close(figure)
