+++ src/nefas/runner.py
@@ -5,8 +5,8 @@ import sys
 from pathlib import Path
 
 from .config import SimulationConfig, load_config
+from .engine import run_simulation
 from .preprocessing import prepare_intermediates
-from .rainfall_simulation import run_rainfall_only_simulation
 
 LOGGER = logging.getLogger(__name__)
 
@@ -21,7 +21,7 @@ def prepare_run(config_path: Path) -> SimulationConfig:
         "load configuration",
         "prepare output directory",
         "prepare geospatial intermediates",
-        "run rainfall-only simulation",
+        "run simulation",
     )
     with tqdm(steps, desc="Preparing run", unit="step", disable=not sys.stderr.isatty()) as progress:
         config = load_config(config_path)
@@ -36,7 +36,7 @@ def prepare_run(config_path: Path) -> SimulationConfig:
         LOGGER.info("Intermediate workspace is ready at %s", intermediates.workspace)
         progress.update()
 
-        simulation = run_rainfall_only_simulation(config, intermediates)
+        simulation = run_simulation(config, intermediates)
         LOGGER.info("Snapshot directory is ready at %s", simulation.snapshot_directory)
         progress.update()
